reorganize testbed app
