+++ examples/babel.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+    plugins: [
+        "@babel/plugin-transform-class-properties",
+        "const-enum",
+        "@babel/plugin-transform-runtime",
+    ],
+    presets: [
+        "@babel/preset-env",
+        "@babel/preset-typescript",
+        "@babel/preset-react",
+    ],
+    env: {
+        es: {
+            presets: [["@babel/preset-env", { modules: false }]],
+        },
+    },
+};
