+++ examples/package.json
@@ -4,11 +4,22 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "serve-example": "webpack serve --config webpack.dev.js",
+    "build-example": "webpack --config webpack.dev.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "detect-collisions": "^9.2.3"
+    "detect-collisions": "^9.2.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "copy-webpack-plugin": "^12.0.2",
+    "css-loader": "^6.9.1",
+    "html-webpack-plugin": "^5.6.0",
+    "mini-css-extract-plugin": "^2.7.7",
+    "webpack": "^5.90.0"
   }
 }
