+++ examples/src/PointSimulatorLive.ts
@@ -1,14 +1,14 @@
 import {
     IClientSimulatorImpl,
     ClientMessageEnum,
-} from "../src/simularium/localSimulators/IClientSimulatorImpl";
+} from "../../src/simularium/localSimulators/IClientSimulatorImpl";
 import {
     EncodedTypeMapping,
     TrajectoryFileInfo,
     VisDataMessage,
-} from "../src/simularium/types";
-import VisTypes from "../src/simularium/VisTypes";
-import { DEFAULT_CAMERA_SPEC } from "../src/constants";
+} from "../../src/simularium/types";
+import VisTypes from "../../src/simularium/VisTypes";
+import { DEFAULT_CAMERA_SPEC } from "../../src/constants";
 
 export default class PointSimLive implements IClientSimulatorImpl {
     nPoints: number;
