+++ examples/src/SingleCurveSimulator.ts
@@ -1,25 +1,23 @@
 import {
     IClientSimulatorImpl,
     ClientMessageEnum,
-} from "../src/simularium/localSimulators/IClientSimulatorImpl";
+} from "../../src/simularium/localSimulators/IClientSimulatorImpl";
 import {
     EncodedTypeMapping,
     TrajectoryFileInfo,
     VisDataMessage,
-} from "../src/simularium/types";
-import VisTypes from "../src/simularium/VisTypes";
-import { DEFAULT_CAMERA_SPEC } from "../src/constants";
+} from "../../src/simularium/types";
+import VisTypes from "../../src/simularium/VisTypes";
+import { DEFAULT_CAMERA_SPEC } from "../../src/constants";
 
-export default class PointSim implements IClientSimulatorImpl {
-    nPoints: number;
-    pointsData: number[];
+export default class SingleCurveSim implements IClientSimulatorImpl {
+    curveData: number[];
+    nPointsPerCurve: number;
     currentFrame: number;
-    nTypes: number;
 
-    constructor(nPoints: number, nTypes: number) {
-        this.nPoints = nPoints;
-        this.nTypes = nTypes;
-        this.pointsData = this.makePoints(nPoints);
+    constructor() {
+        this.nPointsPerCurve = 5;
+        this.curveData = this.makeCurveBundle(1, this.nPointsPerCurve);
         this.currentFrame = 0;
     }
 
@@ -49,52 +47,70 @@ export default class PointSim implements IClientSimulatorImpl {
         ];
     }
 
-    private makePoints(nPoints) {
-        const pts: number[] = [];
-        let p: number[] = [];
-        for (let i = 0; i < nPoints; ++i) {
-            p = this.randomPtInBox(-4, 4, -4, 4, -4, 4);
-            pts.push(p[0]);
-            pts.push(p[1]);
-            pts.push(p[2]);
-        }
-        return pts;
+    private makeCurveBundle() {
+        // make one curve w/5 pts
+        const curves: number[] = [];
+        let p: number[];
+        p = this.randomPtInBox(-4, -3, -2, 2, -2, 2);
+        curves.push(p[0]);
+        curves.push(p[1]);
+        curves.push(p[2]);
+        p = this.randomPtInBox(-2.5, -2, -1, 1, -1, 1);
+        curves.push(p[0]);
+        curves.push(p[1]);
+        curves.push(p[2]);
+        p = this.randomPtInBox(-1, 1, -0.5, 0.5, -0.5, 0.5);
+        curves.push(p[0]);
+        curves.push(p[1]);
+        curves.push(p[2]);
+        p = this.randomPtInBox(2, 2.5, -1, 1, -1, 1);
+        curves.push(p[0]);
+        curves.push(p[1]);
+        curves.push(p[2]);
+        p = this.randomPtInBox(3, 4, -2, 2, -2, 2);
+        curves.push(p[0]);
+        curves.push(p[1]);
+        curves.push(p[2]);
+        return curves;
     }
 
     public update(_dt: number): VisDataMessage {
+        const nFloatsPerCurve = this.nPointsPerCurve * 3;
         //const dt_adjusted = dt / 1000;
         const amplitude = 0.05;
-        for (let ii = 0; ii < this.nPoints; ++ii) {
-            this.pointsData[ii * 3 + 0] += this.randomFloat(
+        for (let jj = 0; jj < this.nPointsPerCurve; ++jj) {
+            this.curveData[jj * 3 + 0] += this.randomFloat(
                 -amplitude,
                 amplitude
             );
-            this.pointsData[ii * 3 + 1] += this.randomFloat(
+            this.curveData[jj * 3 + 1] += this.randomFloat(
                 -amplitude,
                 amplitude
             );
-            this.pointsData[ii * 3 + 2] += this.randomFloat(
+            this.curveData[jj * 3 + 2] += this.randomFloat(
                 -amplitude,
                 amplitude
             );
         }
         // fill agent data.
         const agentData: number[] = [];
-        for (let ii = 0; ii < this.nPoints; ++ii) {
-            agentData.push(VisTypes.ID_VIS_TYPE_DEFAULT); // vis type
-            agentData.push(ii); // instance id
-            agentData.push(ii % this.nTypes); // type
-            agentData.push(this.pointsData[ii * 3 + 0]); // x
-            agentData.push(this.pointsData[ii * 3 + 1]); // y
-            agentData.push(this.pointsData[ii * 3 + 2]); // z
-            agentData.push(0); // rx
-            agentData.push(0); // ry
-            agentData.push(0); // rz
-            agentData.push(0.1); // collision radius
-            agentData.push(0); // subpoints
+        agentData.push(VisTypes.ID_VIS_TYPE_FIBER); // vis type
+        agentData.push(0); // instance id
+        agentData.push(0); // type
+        agentData.push(3.0 * Math.sin(this.currentFrame / 50)); // x
+        agentData.push(0); // y
+        agentData.push(0); // z
+        agentData.push(0); // rx
+        agentData.push(0); // ry
+        agentData.push(this.currentFrame / 50); // rz
+        agentData.push(0.3); // collision radius
+        agentData.push(nFloatsPerCurve);
+        for (let jj = 0; jj < nFloatsPerCurve; ++jj) {
+            agentData.push(this.curveData[jj]);
         }
+
         const frameData: VisDataMessage = {
-            // TODO get msgType out of here
+            // TODO get msgType and connId out of here
             msgType: ClientMessageEnum.ID_VIS_DATA_ARRIVE,
             bundleStart: this.currentFrame,
             bundleSize: 1, // frames
@@ -113,9 +129,7 @@ export default class PointSim implements IClientSimulatorImpl {
 
     public getInfo(): TrajectoryFileInfo {
         const typeMapping: EncodedTypeMapping = {};
-        for (let i = 0; i < this.nTypes; ++i) {
-            typeMapping[i] = { name: `point${i}` };
-        }
+        typeMapping[0] = { name: "fiber0" };
         return {
             // TODO get msgType and connId out of here
             connId: "hello world",
