+++ examples/src/assets/splat.png
