+++ package.json
@@ -6,9 +6,7 @@
     "module": "es/index.js",
     "types": "type-declarations",
     "scripts": {
-        "serve-example": "webpack serve --config webpack.dev.js",
-        "build-example": "webpack --config webpack.dev.js",
-        "start": "npm-run-all build serve-example",
+        "start": "npm run build && cd examples && npm run serve-example",
         "gh-build": "npm-run-all processCSS build-example",
         "lint": "eslint src --ext .ts,.tsx,.js,.jsx",
         "test": "jest --coverage",
@@ -60,20 +58,16 @@
         "babel-loader": "~8.2.2",
         "babel-plugin-const-enum": "^1.1.0",
         "blob-polyfill": "^6.0.20211015",
-        "copy-webpack-plugin": "~9.0.1",
         "copyfiles": "^2.4.1",
         "cross-env": "7.0.3",
-        "css-loader": "^6.7.3",
         "esbuild": "^0.12.25",
         "esbuild-jest": "^0.5.0",
         "eslint": "^7.19.0",
         "eslint-config-prettier": "^6.11.0",
         "eslint-plugin-react": "^7.22.0",
-        "html-webpack-plugin": "^5.3.2",
         "husky": "^1.3.1",
         "jest": "^27.1.0",
         "lint-staged": "^10.5.4",
-        "mini-css-extract-plugin": "^2.1.0",
         "npm-run-all": "^4.1.5",
         "nyc": "~15.0",
         "postcss": "^8.2.10",
@@ -105,7 +99,6 @@
         "@tweakpane/core": "^1.0.8",
         "@tweakpane/plugin-essentials": "^0.1.4",
         "comlink": "^4.3.1",
-        "detect-collisions": "^9.2.2",
         "js-logger": "^1.6.1",
         "parse-pdb": "^1.0.0",
         "si-prefix": "^0.2.0",
